Rewrite: URIs, external graph & incremental build
Use full URIs as object/process identifiers, so that we can link to
their definitions using intersphinx.

External RDF data with definitions can be read, and used to make
cross-references.

This requires that the details of processes and objects are added later
in the building process (later than when the directives are read) so
placeholder docutils elements have been added which are resolved in a
post-transform.

Attempt also to properly clear old data so that we can rebuild
incrementally.